--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/libp2p/go-eventbus
 
 go 1.12
+
+require github.com/jbenet/go-detect-race v0.0.0-20150302022421-3463798d9574