--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,7 @@ module github.com/libp2p/go-eventbus
 
 go 1.12
 
-require github.com/jbenet/go-detect-race v0.0.0-20150302022421-3463798d9574
+require (
+	github.com/jbenet/go-detect-race v0.0.0-20150302022421-3463798d9574
+	github.com/libp2p/go-libp2p-core v0.0.4-0.20190619144953-3605e610067d
+)