For https://github.com/libp2p/go-libp2p/issues/653

TODO:
* [x] Nice README
* [ ] More Docs
* [ ] Catch-all / Send-any - to allow bridging buses (also over the network - might be useful for libp2p daemon) - Can probably leave in an issue
* [ ] CI

Currently I'm getting about 500k-1000k events/s on my machine, optimizing this further would probably be a bit overkill

Benchmark results:
```
goos: linux
goarch: amd64
pkg: github.com/libp2p/go-eventbus
BenchmarkSubs-32         	    2000	    558536 ns/op	     873 B/op	       7 allocs/op
BenchmarkEmits-32        	    2000	    744365 ns/op	     358 B/op	       6 allocs/op
BenchmarkMsgs-32         	    2000	    947316 ns/op	      22 B/op	       0 allocs/op
BenchmarkOneToMany-32    	   20000	     87419 ns/op	     592 B/op	       5 allocs/op
BenchmarkManyToOne-32    	   10000	    129909 ns/op	     191 B/op	       5 allocs/op
BenchmarkMs1e2m4-32      	 1000000	      2148 ns/op	       0 B/op	       0 allocs/op
BenchmarkMs1e0m6-32      	 1000000	      3522 ns/op	       0 B/op	       0 allocs/op // no allocations when sending events
BenchmarkMs0e6m0-32      	 1000000	      4716 ns/op	     698 B/op	       7 allocs/op
BenchmarkMs6e0m0-32      	 1000000	      4785 ns/op	     447 B/op	       6 allocs/op
PASS
```