@raulk 2019-06-14
Particularly, the 1:1 channel/event association creates a DX different that I had in mind, and it makes it impossible to consume logically related events, or events with _happens-before_ relations (eg. stream open/close, etc.) in the order they were emitted, or more strictly speaking, pushed into the receiver.