@magik6k 2019-06-16
I've changed the interface to now use user-provided typed channels as discussed in the issue. Perf impact is quite minimal for the rather improved DX experience:

```
benchmark                 old ns/op     new ns/op     delta
BenchmarkSubs-32          509731        600981        +17.90%
BenchmarkEmits-32         998095        622095        -37.67%
BenchmarkMsgs-32          931380        888318        -4.62%
BenchmarkOneToMany-32     82220         83740         +1.85%
BenchmarkManyToOne-32     126449        127280        +0.66%
BenchmarkMs1e2m4-32       2053          1911          -6.92%
BenchmarkMs1e0m6-32       317           353           +11.36% // probably most important, sending 1M messages to 10 subs
BenchmarkMs0e6m0-32       4765          4720          -0.94%
BenchmarkMs6e0m0-32       4394          4264          -2.96%

benchmark                 old allocs     new allocs     delta
BenchmarkSubs-32          5              4              -20.00%
BenchmarkEmits-32         5              6              +20.00%
BenchmarkMsgs-32          0              0              +0.00%
BenchmarkOneToMany-32     5              4              -20.00%
BenchmarkManyToOne-32     5              5              +0.00%
BenchmarkMs1e2m4-32       0              0              +0.00%
BenchmarkMs1e0m6-32       0              0              +0.00%
BenchmarkMs0e6m0-32       7              7              +0.00%
BenchmarkMs6e0m0-32       6              5              -16.67%

benchmark                 old bytes     new bytes     delta
BenchmarkSubs-32          394           415           +5.33%
BenchmarkEmits-32         217           256           +17.97%
BenchmarkMsgs-32          24            26            +8.33%
BenchmarkOneToMany-32     382           417           +9.16%
BenchmarkManyToOne-32     206           205           -0.49%
BenchmarkMs1e2m4-32       0             0             +0.00%
BenchmarkMs1e0m6-32       0             0             +0.00%
BenchmarkMs0e6m0-32       632           657           +3.96%
BenchmarkMs6e0m0-32       473           505           +6.77%
```