@Kubuxu 2019-06-19
I think we should keep this as simple as possible. Bridging over the network should not be within the feature set as you start running into issues with reliable delivery, one and only one delivery and so on.

I also think we shouldn't concern ourselves with things like `keepLast`. The assumption I would make is that event bus and subscribers are set up before the services start publishing.
The keeplast doesn't help you anyway in that case.