@magik6k 2019-06-19
So one of the use cases for the eventbus might be to eventually allow reconfiguring the node in runtime, which may eventually need the flexibility this option gives.

Also, the overhead of it is quite low:
```
BenchmarkMs1e0m6-32             10000000               513 ns/op               0 B/op          0 allocs/op
BenchmarkMs0e0m6-32              1000000               227 ns/op               0 B/op          0 allocs/op
BenchmarkStatefulMs1e0m6-32     10000000               448 ns/op               3 B/op          0 allocs/op
BenchmarkStatefulMs0e0m6-32      1000000               304 ns/op              32 B/op          1 allocs/op
```

We can either just not include it (it will be easy to but it back when we need it again), or just not use.