@magik6k 2019-06-19
benchcmp for head vs head~1:
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkSubs-32                553005        536262        -3.03%
BenchmarkEmits-32               1290763       1333304       +3.30%
BenchmarkMsgs-32                997285        1028917       +3.17%
BenchmarkOneToMany-32           79473         82656         +4.01%
BenchmarkManyToOne-32           129494        126470        -2.34%
BenchmarkMs1e2m4-32             2067          2116          +2.37%
BenchmarkMs1e0m6-32             485           444           -8.45%
BenchmarkMs0e0m6-32             217           216           -0.46%
BenchmarkStatefulMs1e0m6-32     453           436           -3.75%
BenchmarkStatefulMs0e0m6-32     298           293           -1.68%
BenchmarkMs0e6m0-32             5169          4856          -6.06%
BenchmarkMs6e0m0-32             4418          4947          +11.97%

benchmark                       old allocs     new allocs     delta
BenchmarkSubs-32                5              5              +0.00%
BenchmarkEmits-32               7              7              +0.00%
BenchmarkMsgs-32                0              0              +0.00%
BenchmarkOneToMany-32           5              5              +0.00%
BenchmarkManyToOne-32           7              6              -14.29%
BenchmarkMs1e2m4-32             0              0              +0.00%
BenchmarkMs1e0m6-32             0              0              +0.00%
BenchmarkMs0e0m6-32             0              0              +0.00%
BenchmarkStatefulMs1e0m6-32     0              0              +0.00%
BenchmarkStatefulMs0e0m6-32     1              1              +0.00%
BenchmarkMs0e6m0-32             8              7              -12.50%
BenchmarkMs6e0m0-32             6              6              +0.00%

benchmark                       old bytes     new bytes     delta
BenchmarkSubs-32                413           378           -8.47%
BenchmarkEmits-32               204           176           -13.73%
BenchmarkMsgs-32                24            25            +4.17%
BenchmarkOneToMany-32           434           431           -0.69%
BenchmarkManyToOne-32           189           152           -19.58%
BenchmarkMs1e2m4-32             0             0             +0.00%
BenchmarkMs1e0m6-32             0             0             +0.00%
BenchmarkMs0e0m6-32             0             0             +0.00%
BenchmarkStatefulMs1e0m6-32     3             3             +0.00%
BenchmarkStatefulMs0e0m6-32     32            32            +0.00%
BenchmarkMs0e6m0-32             587           559           -4.77%
BenchmarkMs6e0m0-32             503           499           -0.80%

```