@raulk 2019-06-19
Oh, we need to merge the PR in go-libp2p-core, make a release, and update go.mod here.