@magik6k 2019-06-19
Ran the benchmarks comparing between last 3 commits, it's somehow a bit faster:

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkSubs-32                587355        512941        -12.67%
BenchmarkEmits-32               1294642       1070025       -17.35%
BenchmarkMsgs-32                1023572       960473        -6.16%
BenchmarkOneToMany-32           81675         81233         -0.54%
BenchmarkManyToOne-32           130050        121522        -6.56%
BenchmarkMs1e2m4-32             2167          2081          -3.97%
BenchmarkMs1e0m6-32             531           358           -32.58%
BenchmarkMs0e0m6-32             227           217           -4.41%
BenchmarkStatefulMs1e0m6-32     465           449           -3.44%
BenchmarkStatefulMs0e0m6-32     303           289           -4.62%
BenchmarkMs0e6m0-32             5146          4727          -8.14%
BenchmarkMs6e0m0-32             4842          4859          +0.35%

benchmark                       old allocs     new allocs     delta
BenchmarkSubs-32                5              5              +0.00%
BenchmarkEmits-32               4              7              +75.00%
BenchmarkMsgs-32                0              0              +0.00%
BenchmarkOneToMany-32           5              5              +0.00%
BenchmarkManyToOne-32           4              6              +50.00%
BenchmarkMs1e2m4-32             0              0              +0.00%
BenchmarkMs1e0m6-32             0              0              +0.00%
BenchmarkMs0e0m6-32             0              0              +0.00%
BenchmarkStatefulMs1e0m6-32     0              0              +0.00%
BenchmarkStatefulMs0e0m6-32     1              1              +0.00%
BenchmarkMs0e6m0-32             5              7              +40.00%
BenchmarkMs6e0m0-32             6              6              +0.00%

benchmark                       old bytes     new bytes     delta
BenchmarkSubs-32                362           419           +15.75%
BenchmarkEmits-32               148           222           +50.00%
BenchmarkMsgs-32                41            21            -48.78%
BenchmarkOneToMany-32           431           433           +0.46%
BenchmarkManyToOne-32           130           176           +35.38%
BenchmarkMs1e2m4-32             0             0             +0.00%
BenchmarkMs1e0m6-32             0             0             +0.00%
BenchmarkMs0e0m6-32             0             0             +0.00%
BenchmarkStatefulMs1e0m6-32     3             3             +0.00%
BenchmarkStatefulMs0e0m6-32     32            32            +0.00%
BenchmarkMs0e6m0-32             536           572           +6.72%
BenchmarkMs6e0m0-32             498           501           +0.60%
```

The dep list is a bit painful, but afaik gomod doesn't download things that aren't being imported so it's probably fine.

Code LGTM

:ship: